split NearVmMemory into a separate file to limit visibility
